# Conflicts:
#	.idea/caches/build_file_checksums.ser
#	.idea/misc.xml
#	app/src/main/java/hft/wiinf/de/horario/model/Event.java
#	app/src/main/java/hft/wiinf/de/horario/view/CalendarActivity.java

+++ app/src/main/java/hft/wiinf/de/horario/controller/EventController.java
@@ -9,18 +9,20 @@ import java.util.Date;
 import java.util.GregorianCalendar;
 import java.util.List;
 
+import hft.wiinf.de.horario.model.AcceptedState;
 import hft.wiinf.de.horario.model.Event;
 import hft.wiinf.de.horario.model.Person;
 import hft.wiinf.de.horario.model.Repetition;
 
 public class EventController {
     //saves (update or create)an event
-    public static void saveEvent(Event event) {
+    public static void saveEvent(@NonNull Event event) {
         event.save();
     }
 
-    public static void deleteEvent(Event event) {
+    public static void deleteEvent(@NonNull Event event) {
         //deletes all persons that accepted the event
+
         for (Person person : PersonController.getEventCancelledPersons(event)) {
             PersonController.deletePerson(person);
         }
@@ -29,14 +31,19 @@ public class EventController {
         for (Person person : PersonController.getEventAcceptedPersons(event)) {
             PersonController.deletePerson(person);
         }
-        event.delete();
+        if (EventController.findRepeatingEvents(event.getId()).size() > 0) {
+            event.setAccepted(AcceptedState.REJECTED);
+            event.save();
+        } else {
+            event.delete();
+        }
     }
 
     public static Event getEventById(@NonNull Long id) {
         return Event.load(Event.class, id);
     }
 
-    //find the list of events that start in the given period (enddate is ecluded!)
+    //find the list of events that start in the given period (enddate is not included!)
     public static List<Event> findEventsByTimePeriod(Date startDate, Date endDate) {
         return new Select().from(Event.class).where("starttime between ? AND ?", startDate.getTime(), endDate.getTime() - 1).execute();
 
@@ -45,6 +52,10 @@ public class EventController {
     //get a list of all events that I accepted
     public static List<Event> findMyAcceptedEvents() {
         return new Select().from(Event.class).where("accepted=?", true).execute();
+    }
+
+    public static List<Event> findRepeatingEvents(@NonNull Long eventId) {
+        return new Select().from(Event.class).where("startevent=?", eventId).execute();
     }
 
     // saves a serial event, firstEvent="StartEvent", repetition: repetition frequence (daily, ...), endOfRepetiton: last day of the repetition (including)
@@ -71,10 +82,13 @@ public class EventController {
             Event repetitionEvent = new Event();
             repetitionEvent.setPlace(firstEvent.getPlace());
             repetitionEvent.setDescription(firstEvent.getDescription());
-            repetitionEvent.setAccepted(firstEvent.isAccepted());
+            repetitionEvent.setAccepted(firstEvent.getAccepted());
             repetitionEvent.setCreator(firstEvent.getCreator());
             repetitionEvent.setStartTime(firstEvent.getStartTime());
             repetitionEvent.setEndTime(firstEvent.getEndTime());
+            repetitionEvent.setEndTime(firstEvent.getEndDate());
+            repetitionEvent.setShortTitle(firstEvent.getShortTitle());
+            repetitionEvent.setStartEvent(firstEvent);
             Calendar temporary = new GregorianCalendar();
             temporary.setTime(repetitionEvent.getStartTime());
             temporary.add(fieldNumber, i);
@@ -82,7 +96,7 @@ public class EventController {
             temporary.setTime(repetitionEvent.getEndTime());
             temporary.add(fieldNumber, i);
             repetitionEvent.setEndTime(temporary.getTime());
-            //if end of repetition is overruned, stopp eelse save the new Event;
+            //if end of repetition is overruned, stop,p else save the new Event;
             if (repetitionEvent.getStartTime().after(endOfRepetition.getTime()))
                 break;
             saveEvent(repetitionEvent);
